hackathon_projects/women_safety_app/backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
bcrypt==5.0.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cryptography==46.0.2
ecdsa==0.19.1
fastapi==0.118.2
frozenlist==1.8.0
h11==0.16.0
idna==3.10
multidict==6.7.0
passlib==1.7.4
propcache==0.3.2
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.0
pydantic_core==2.41.1
PyJWT==2.10.1
PyMySQL==1.1.2
python-dotenv==1.1.1
python-jose==3.5.0
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
twilio==9.8.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
yarl==1.22.0
